separate demangler into support library
